modified:   src/app-src/components/ProgressBar.jsx 	modified:   src/app-src/components/VolumeBar.jsx 	modified:   src/app-src/layout/layout-content/PlayBar.jsx 	modified:   src/pages/Content.jsx 	modified:   src/pages/MyPlaylist.jsx 	modified:   src/pages/PageLayout.jsx 	modified:   src/store/slice/musicSlice.js
